chore: add types for method params

+++ src/repositories/transactions-repository.ts
@@ -1,17 +1,41 @@
 import { Transaction } from "./model/transtaction";
 
+export interface CreateTransactionParams {
+  title: string;
+  description: string;
+  value: number;
+  user_id: string;
+}
+
+export interface FindManyTransactionsParams {
+  user_id: string;
+  title: string;
+  page: number;
+  from?: string;
+  to?: string;
+  order?: "desc" | "asc";
+}
+
+export interface FindByIdTransactionParams {
+  user_id: string;
+  id: string;
+}
+
+export interface DeleteByIdTransactionParams extends FindByIdTransactionParams {
+ 
+}
+
+export interface UpdateByIdTransactionParams extends CreateTransactionParams {
+  id: string;
+}
+
 export interface TransactionsRepository {
   create: ({
     title,
     description,
     value,
     user_id,
-  }: {
-    title: string;
-    description: string;
-    value: number;
-    user_id: string;
-  }) => Promise<Transaction>;
+  }: CreateTransactionParams) => Promise<Transaction>;
 
   findMany: ({
     user_id,
@@ -20,42 +44,20 @@ export interface TransactionsRepository {
     from,
     to,
     order,
-  }: {
-    user_id: string;
-    title: string;
-    page: number;
-    from: string | null;
-    to: string | null;
-    order: "desc" | "asc" | null;
-  }) => Promise<Transaction[]>;
+  }: FindManyTransactionsParams) => Promise<Transaction[]>;
 
   findById: ({
     user_id,
-    transaction_id,
-  }: {
-    user_id: string;
-    transaction_id: string;
-  }) => Promise<Transaction | null>;
+    id,
+  }: FindByIdTransactionParams) => Promise<Transaction | null>;
 
-  deleteById: ({
-    user_id,
-    transaction_id,
-  }: {
-    user_id: string;
-    transaction_id: string;
-  }) => Promise<string>;
+  deleteById: ({ user_id, id }: DeleteByIdTransactionParams) => Promise<string>;
 
   updateById: ({
     user_id,
-    transaction_id,
+    id,
     title,
     description,
     value,
-  }: {
-    user_id: string;
-    transaction_id: string;
-    title: string;
-    description: string;
-    value: number;
-  }) => Promise<Transaction>;
+  }: UpdateByIdTransactionParams) => Promise<Transaction>;
 }
